Added guide for video upload issues (#79)
* Added guide for video upload issues

As I'm not well versed in the upload part of the site, I tried to provide as much information as I know and have seen other people suggest, such as reuploading.

I also added a section for helping people use ffprobe to see if they have a video stream encoding error.

Open to suggestions for additions, or if it's even worth merging.

- A.S.

* Apply suggestions from code review

Co-authored-by: Russ Smith <russ@bashme.org>

---------

Co-authored-by: Jeremy Woertink <jeremywoertink@gmail.com>
Co-authored-by: Russ Smith <russ@bashme.org>